Clean up some nitpicks.
